% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_each_start_parallel.R
\name{for_each_start_parallel}
\alias{for_each_start_parallel}
\title{Start parallel computing on non-Windows}
\usage{
for_each_start_parallel(cores)
}
\arguments{
\item{cores}{Number of cores to use}
}
\description{
Start parallel computing on non-Windows
}
