% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_each_start_parallel_windows.R
\name{for_each_start_parallel_windows}
\alias{for_each_start_parallel_windows}
\title{Start parallel computing on Windows}
\usage{
for_each_start_parallel_windows(cores)
}
\arguments{
\item{cores}{Number of cores to use}
}
\value{
A PSOCK cluster object
}
\description{
Start parallel computing on Windows
}
