% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_artificial_dataset.R
\name{generate_artificial_dataset}
\alias{generate_artificial_dataset}
\title{Artificial Data Set}
\usage{
generate_artificial_dataset(N)
}
\arguments{
\item{N}{Sample size}
}
\value{
The artificial data set
}
\description{
Generate an artificial dataset
}
