% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_csv_in_temp_folder.R
\name{generate_csv_in_temp_folder}
\alias{generate_csv_in_temp_folder}
\title{Generate a .csv file in a temp folder}
\usage{
generate_csv_in_temp_folder(tbl, file_name = "temp")
}
\arguments{
\item{tbl}{the data set of interest.}

\item{file_name}{name of the temp file; Default "temp".}
}
\value{
A .csv file in the temp folder.
}
\description{
Create a temp folder in your project directory
and generate a csv file based on your data set of interest,
which can be conveniently pasted to other reports.
}
