% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart_pie_add_percentage.R
\name{hchart_pie_add_percentage}
\alias{hchart_pie_add_percentage}
\title{Add percentage label to a highchart pie chart}
\usage{
hchart_pie_add_percentage(.hchart_pie)
}
\arguments{
\item{.hchart_pie}{A highchart pie chart object.}
}
\value{
A highchart pie chart object with percentage labels added.
}
\description{
Add percentage label to a highchart pie chart
}
