% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_cores.R
\name{num_cores}
\alias{num_cores}
\title{Report number of cores}
\usage{
num_cores(logical = FALSE)
}
\arguments{
\item{logical}{if logical cores}
}
\value{
Number of cores
}
\description{
Report number of cores
}
