% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nas_with_zeros.R
\name{replace_nas_with_zeros}
\alias{replace_nas_with_zeros}
\title{Replace NAs with zeros}
\usage{
replace_nas_with_zeros(tbl)
}
\arguments{
\item{tbl}{A data frame to fill.}
}
\value{
The filled data frame.
}
\description{
Replace NAs with zeros
}
