% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_col_names.R
\name{str_col_names}
\alias{str_col_names}
\title{Extract column names with regular expressions}
\usage{
str_col_names(data, pattern)
}
\arguments{
\item{data}{A data frame.}

\item{pattern}{A pattern.}
}
\value{
Column names that match the pattern.
}
\description{
Extract column names with regular expressions
}
\examples{
str_col_names(iris, "Length")
}
