% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_states.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{US states info}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 7 columns.
}
\usage{
us_states
}
\description{
US states info
}
\keyword{datasets}
